backend/requirements.txt
Django
djangorestframework
django-cors-headers
python-dotenv
coreapi
setuptools